+++ Tests/DeserializationTests.cs
@@ -2,7 +2,7 @@
 using Newtonsoft.Json;
 using System.IO;
 using System.Linq;
-using DeserializedTable = Timetable.Deserialized.Table;
+using PersistenceTable = Timetable.Persistence.Table;
 using Timetable;
 
 namespace Tests
@@ -13,38 +13,38 @@ namespace Tests
         [TestMethod]
         public void Deserialized_table_should_not_be_null()
         {
-            var table = Deserialize("TimeTable.json");
+            var table = Load("TimeTable.json");
             Assert.IsNotNull(table, "table should not be null!");
         }
 
         [TestMethod]
         public void Deserialized_block_start_time_should_be_available()
         {
-            var table = Deserialize("TimeTable.json");
+            var table = Load("TimeTable.json");
             Assert.IsNotNull(table.StartTimes, "Block start times should not be null!");
         }
 
         [TestMethod]
         public void Deserialized_block_should_contain_subject_name()
         {
-            var table = Deserialize("TimeTable.json");
+            var table = Load("TimeTable.json");
             var block = table.Weekdays.First().Blocks.First();
             Assert.IsFalse(string.IsNullOrEmpty(block.Name), "Block should contain subject name!");
         }
 
         [TestMethod]
-        public void Converting_deserialized_table_to_regular_table_should_succeed()
+        public void Convert_persistence_table_to_regular_table_should_succeed()
         {
-            var deserialized = Deserialize("TimeTable.json");
-            var table = TableConvert.FromDeserialized(deserialized);
-            Assert.IsNotNull(table, "Table should not be null!");
+            var table = Load("TimeTable.json");
+            var regular = TableConvert.FromPersistence(table);
+            Assert.IsNotNull(regular, "Regular table should not be null!");
         }
 
-        private DeserializedTable Deserialize(string fileName)
+        private PersistenceTable Load(string fileName)
         {
             var path = Path.Combine("Data", fileName);
             var json = File.ReadAllText(path);
-            return JsonConvert.DeserializeObject<DeserializedTable>(json);
+            return JsonConvert.DeserializeObject<PersistenceTable>(json);
         }
     }
 }
